# webroot/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cakephp Api",
        "description": "Example swagger openApi",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http:\/\/localhost:8765"
        }
    ],
    "paths": {
        "\/api\/episodes\/{programId}.json": {
            "get": {
                "operationId": "3832575498952b92e073d46a4a8ed12d",
                "parameters": [
                    {
                        "name": "programId",
                        "in": "path",
                        "description": "Program ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list episodes of program ID",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/EpisodesTable"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/episodes.json": {
            "post": {
                "operationId": "a233cb3ab7d3f2d196f8332773eb35de",
                "parameters": [
                    {
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "description": "CSRF token for security",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/FormEpisode"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "create episode",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/EpisodesTable"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/demo.json": {
            "get": {
                "operationId": "7898612985d0f71dee4cfaef4d3514b8",
                "responses": {
                    "200": {
                        "description": "List demo"
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/api\/csrf.json": {
            "get": {
                "operationId": "24846a7632cc8d2deb2c6d252db79218",
                "responses": {
                    "200": {
                        "description": "Csrf token"
                    },
                    "401": {
                        "description": "Not allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "FormEpisode": {
                "properties": {
                    "title": {
                        "description": "Episode Entity",
                        "type": "string",
                        "example": "My Episode Title"
                    },
                    "slug": {
                        "type": "string",
                        "example": "my-episode-title"
                    },
                    "youtube": {
                        "type": "string",
                        "example": "https:\/\/www.youtube.com\/watch?v=xxxxxxx",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "A short description of the episode."
                    },
                    "online": {
                        "type": "integer",
                        "example": 1
                    },
                    "program_id": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            },
            "EpisodesTable": {
                "type": "array",
                "items": {
                    "properties": {
                        "id": {
                            "description": "Identifiant unique de l'episode",
                            "type": "integer"
                        },
                        "title": {
                            "description": "title episode",
                            "type": "string"
                        },
                        "slug": {
                            "description": "slug episode",
                            "type": "string"
                        },
                        "youtube id": {
                            "description": "youtube Id",
                            "type": "string"
                        },
                        "description": {
                            "description": "description episode",
                            "type": "string"
                        },
                        "online": {
                            "description": "online episode",
                            "type": "integer"
                        },
                        "created_at": {
                            "description": "yyyy-mm-dd hh:ii:ss category created",
                            "type": "string",
                            "format": "date-time"
                        },
                        "updated_at": {
                            "description": "yyyy-mm-dd hh:ii:ss category updated",
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    "type": "object"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}